chore : Rename route and Seperate category route
